feat : 구매한 로또 상태를 표현하기 위한 상태관리 클래스 추가
각각 로또 추첨에 필요한 데이터들과 사용자에게 출력할 message를 인자로 갖고 있습니다.
