+++ src/main/kotlin/lotto/presentation/vm/model/PurchaseInfo.kt
@@ -1,0 +1,10 @@
+package lotto.presentation.vm.model
+
+import lotto.domain.enums.Output.Companion.purchaseFormat
+
+data class PurchaseInfo(
+    val purchaseLottoCount: Int = 0,
+){
+    val guideMessage: String
+        get() = purchaseFormat(purchaseLottoCount)
+}
