+++ src/main/kotlin/lotto/presentation/vm/model/PurchaseState.kt
@@ -1,19 +1,10 @@
 package lotto.presentation.vm.model
 
-import lotto.domain.enums.Rank
-import java.util.TreeSet
-
 data class PurchaseState(
-    val purchaseLottoCount: Int = 0,
+    val purchaseInfo: PurchaseInfo = PurchaseInfo(),
     val winningNumber: List<Int> = emptyList(),
     val bonusNumber: Int = 0,
-    val pickedLotto: List<TreeSet<Int>> = emptyList(),
-    val winningResult: Map<Rank, Int> = mapOf(
-        Rank.FIFTH to 0,
-        Rank.FOURTH to 0,
-        Rank.THIRD to 0,
-        Rank.SECOND to 0,
-        Rank.FIRST to 0,
-    ),
-    val rateOfReturn: String = "0.0"
+    val lotto: PurchasedLottoInfo = PurchasedLottoInfo(),
+    val winningResult: WinningResult = WinningResult(),
+    val rateOfReturn: String = "",
 )
